chore: move audit models
