+++ apps/ai-dial-admin/src/components/ActivityAudit/Rollback/SystemRollback.tsx
@@ -22,7 +22,7 @@ import Tabs from '@/src/components/Common/Tabs/Tabs';
 import { ActivityAuditI18nKey, MenuI18nKey } from '@/src/constants/i18n';
 import { BASE_ICON_PROPS } from '@/src/constants/main-layout';
 import { useI18n } from '@/src/locales/client';
-import { ActivityAuditDiff, DialActivity } from '@/src/models/dial/activity-audit';
+import { ActivityAuditDiff, DialActivity } from '@/src/models/activity-audit';
 import { EntitiesGridData } from '@/src/models/entities-grid-data';
 import { ActivityAuditEntity, ActivityAuditResourceType, DiffView } from '@/src/types/activity-audit';
 import { PopUpState } from '@/src/types/pop-up';
