+++ apps/ai-dial-admin/src/components/ActivityAudit/View/DiffReport/EntityDiff.tsx
@@ -1,7 +1,7 @@
 import { FC } from 'react';
 
 import { createSectionFromDiffs } from '@/src/components/ActivityAudit/View/utils';
-import { ActivityAuditDiff } from '@/src/models/dial/activity-audit';
+import { ActivityAuditDiff } from '@/src/models/activity-audit';
 import { ActivityAuditResourceType, CompareView, DiffView } from '@/src/types/activity-audit';
 import DiffLegend from '@/src/components/ActivityAudit/View/DiffReport/DiffLegend';
 import DiffSection from '@/src/components/ActivityAudit/View/DiffReport/DiffSection';
