+++ apps/ai-dial-admin/src/components/ActivityAudit/View/Header/Header.tsx
@@ -8,7 +8,7 @@ import { getFormattedResourceType } from '@/src/constants/grid-columns/formatter
 import { ActivityAuditI18nKey } from '@/src/constants/i18n';
 import { BASE_ICON_PROPS } from '@/src/constants/main-layout';
 import { useCurrentLocale, useI18n } from '@/src/locales/client';
-import { DialActivity } from '@/src/models/dial/activity-audit';
+import { DialActivity } from '@/src/models/activity-audit';
 import { ActivityAuditType } from '@/src/types/activity-audit';
 import { ApplicationRoute } from '@/src/types/routes';
 import { formatDateTimeToLocalString } from '@/src/utils/formatting/date';
